No longer rely on GraphQL service for namespace icon
